requirements.txt
aiogram==3.4.1
flask==3.0.3
httpx==0.27.0
openai==1.23.6
python-dotenv==1.0.1
